Se solucionan errores

+++ combo/src/main/java/com/example/combo/combo/repositories/RolRepository.java
@@ -3,6 +3,6 @@ package com.example.combo.combo.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.combo.combo.entities.Rol;
-public interface RolRepository extends JpaRepository<Rol, Long>{
+public interface RolRepository extends JpaRepository<Rol, Integer>{
 
 }
